viewer, editor: Various robustness and UX improvements.
* viewer.c: Fix was_unicode calculation to avoid spurious aborts.
* Reduce FETCHLIST interval, since even an interval of 180
  can take up to 15 seconds to load a folder.
* SMTP: Don't send addresses with double <<>> for RCPT.
* editor: Add spaces at beginning of quoted lines, as with viewer, for
  readability.
* editor: Fix erroneous extra spaces at end of format=flowed lines,
  caused by duplicate logic that would add a space in this case.
* editor: Distinguish between message not sent and sent but not saved.
* imap: Avoid infinite loop by breaking early if flushed input is empty.
* IMAP APPEND: Fix memory leak of flags.

+++ evergreen.c
@@ -1238,7 +1238,7 @@ static void adjust_start_seqno(struct client *client, uint32_t current, int offs
 		/* Going up */
 		uint32_t max_start_seqno;
 
-		if (client->sel_mbox->total > FETCHLIST_INTERVAL) {
+		if (client->sel_mbox->total > (uint32_t) FETCHLIST_INTERVAL) {
 			max_start_seqno = client->sel_mbox->total - FETCHLIST_INTERVAL;
 		} else {
 			max_start_seqno = 1;
